Update German channels (#1)
* Add new links for: ZDF, 3sat, ARTE, ONE, ZDFinfo, ZDFneo
* Remove links for: RTL II and RTLplus, those links are not Germanm, but
Hungarian (.hu); in Germany there are chanels with the same name, but
they are not free
* Remove non-free channels: RTL, SAT.1, ProSieben, VOX, kabel eins,
sixx, DMAX, TELE 5 SAT.1 Gold, ProSieben MAXX, NITRO (those channels are
private channels and according to ubuntuusers.de (see below) are only
accesible through a webbrowser after (free) registration.)
* I've used links in HD quality and added "HD" to the channel names,
because the channels show the "HD" besides their logos as well.
* Changes link for "ARD" = "Das Erste", because the current link seems
to be VERY slow to load, when you switch the channels. The new link
seems to be a little faster.

Used sources:
* https://wiki.ubuntuusers.de/Internet-TV/Stationen
* https://github.com/jnk22/kodinerds-iptv/blob/master/iptv/clean/clean_tv_main.m3u
* https://github.com/jnk22/kodinerds-iptv/blob/master/iptv/clean/clean_tv.m3u

According to the (German) README "Kodinerds IPTV" is a collection of free-to-air
streams for TV and radio stations.