playlist_japan.m3u8
The jp-primehome.com link is a distribution server service that can only be viewed normally by viewers who have a paid viewing contract with the distributor.

The distributor has restricted access to the jp-primehome.com link URL previously posted on GitHub “FreeTV/IPTV,” so it is no longer possible to view it for free as before. Therefore, even if the link continues to be posted, it is unlikely that it will be possible to view it.